requirements.txt
customtkinter
fpdf2

# API server
Flask
flask-cors
requests
waitress
gunicorn
